[feat]tcp socket option

+++ comm/processor/tcp/transmitter.go
@@ -12,7 +12,7 @@ type TCPMessageTransmitter struct {
 }
 
 type socketOpt interface {
-	MaxPacketSize() uint16
+	MaxPacketSize() int
 	ApplySocketReadTimeout(conn net.Conn, cb func())
 	ApplySocketWriteTimeout(conn net.Conn, cb func())
 }
